+++ lib/data/include/metashell/data/wave_arg_parser.hpp
@@ -43,7 +43,6 @@ namespace metashell
       wave_arg_parser& operator=(wave_arg_parser&&) = delete;
 
       void parse(const command_line_argument_list& args_,
-                 include_config sysincludes_,
                  std::vector<std::string> sysmacros_);
 
       const wave_config& result() const;
@@ -53,7 +52,6 @@ namespace metashell
     private:
       data::include_config _includes;
       std::vector<command_line_argument> _standards;
-      bool _use_stdincpp = true;
 
       wave_config _config;