+++ lib/engine/wave/include/metashell/engine/wave/trace.hpp
@@ -37,7 +37,8 @@ namespace metashell
         trace(const data::cpp_code& env_,
               const boost::optional<data::cpp_code>& exp_,
               const data::wave_config& config_,
-              data::metaprogram_mode mode_);
+              data::metaprogram_mode mode_,
+              const std::vector<boost::filesystem::path>& system_includes_);
 
         boost::optional<data::event_data> next();